Naprawiono drobne błędy przy dodawaniu nowego rodzaju stanowiska, dodano opcje zwalniania pracownika.

+++ src/main/java/com/WorkersDataBase/service/position/PositionService.java
@@ -2,7 +2,7 @@ package com.WorkersDataBase.service.position;
 
 import com.WorkersDataBase.data.position.Position;
 import com.WorkersDataBase.data.position.PositionRepository;
-import com.WorkersDataBase.service.Notification.ServicePushNotification;
+import com.WorkersDataBase.service.notification.ServicePushNotification;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,41 +13,56 @@ import java.util.List;
 @AllArgsConstructor
 public class PositionService {
     private final PositionRepository positionRepository;
-    private final ServicePushNotification servicePushNotification;
     private final PositionValidTool positionValidTool;
-    //  addNewPositionType returns true if adding was successful
+    private final ServicePushNotification notification;
+
     @Transactional
     public boolean addNewPositionType(String positionName){
 
-        if (positionValidTool.positionNameIsUnique(positionName)){
-
-            Position position = new Position();
-            position.setPositionName(positionName);
-            positionRepository.save(position);
-
-            servicePushNotification.pushNewPositionSuccess(positionName);
-            return true;
+        //  Valid
+        if(positionValidTool.positionIsNull(positionName)){
+            notification.pushUserTryingAddNullPositionInfo();
+            return false;
+        }
+        if (!positionValidTool.positionNameIsUnique(positionName)){
+            notification.pushUniquePositionError();
+            return false;
+        }
+        if (!positionValidTool.positionNameIsFine(positionName)){
+            notification.pushToShortPositionNameInfo();
+            return false;
         }
 
-        servicePushNotification.pushUniquePositionError();
-        return false;
+        //  Do
+        Position position = new Position();
+        position.setPositionName(positionName);
+        positionRepository.save(position);
+
+        notification.pushNewPositionSuccess(positionName);
+        return true;
+    }
+
+    @Transactional
+    public boolean deletePosition(Position position){
+
+        //  Valid
+        if(positionValidTool.positionIsNull(position)){
+            notification.pushError();
+            return false;
+        }
+
+        if (positionValidTool.someoneHasContract(position)){
+            notification.pushDeletingPositionConflictInfo();
+            return false;
+        }
+        //  Do
+        positionRepository.delete(position);
+        notification.pushDeletePositionSuccess(position);
+        return true;
     }
 
     public List<Position> getPositions(){
         return positionRepository.findAll();
     }
 
-    //  deletePosition returns true if deleting was successful
-    @Transactional
-    public boolean deletePosition(Position position){
-
-        if (positionValidTool.someoneHasContract(position)){
-            servicePushNotification.pushDeletingPositionConflictInfo();
-            return false;
-        }
-
-        positionRepository.delete(position);
-        servicePushNotification.pushDeletePositionSuccess(position);
-        return true;
-    }
 }
